Added a call to RIOT API for the info on all of their items and publishing that to the champ page. I also figured out that the slowest call we make is to get all of the stats about a champ (the game data not the champs info). The best way to speed up the page load is to start off by loading the item info, champion info, and champion list. Then once the frontend js has started executing it can make an ajax call to the backend to get the champion stat data. This way, it looks like it loads quickly because a user will see something sooner.
Therefore, I have added an api endpoint that will return champ data given an id. It is in the format "/api/champ/{id}". If you make a get request to that, you will get back the champ's stat data. Looking for more improvements is my next step. And some cleanup.